verity: Update base::bits::AlignUp arguments
The arguments are required to be unsigned type in CL:5041522

BUG=b:314056626
TEST=FEATURES=test emerge-volteer verity

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5076732
Commit-Queue: Jae Hoon Kim <kimjae@chromium.org>
Tested-by: Byron Lee <byronlee@chromium.org>
Auto-Submit: Byron Lee <byronlee@chromium.org>

+++ verity/file_hasher.cc
@@ -73,7 +73,8 @@ bool FileHasher::Initialize() {
       LOG(ERROR) << "The source file size must be divisible by the block size, "
                  << "Size: " << source_size;
       LOG(INFO) << "Suggested size: "
-                << base::bits::AlignUp(source_size, int64_t{PAGE_SIZE});
+                << base::bits::AlignUp(static_cast<uint64_t>(source_size),
+                                       uint64_t{PAGE_SIZE});
       return false;
     }
   }
